requirements.txt
Flask
Flask-Script
Flask-SQLAlchemy
Flask-Migrate
Flask-Restful
Flask_Cors
Psycopg2
gunicorn
cryptography
PyJWT